chore: refactor DatabaseV2 out of index
