+++ src/stores/Discussions/discussionEvents.ts
@@ -1,7 +1,7 @@
 import { logger } from 'src/logger'
 
 import type { IDiscussion } from 'src/models'
-import type { DatabaseV2 } from '../databaseV2'
+import type { DatabaseV2 } from '../databaseV2/DatabaseV2'
 import type { DBEndpoint } from '../databaseV2/endpoints'
 
 type DiscussionEndpoints = Extract<
